backend/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask_script
pyjwt
flask-swagger-ui
flask_monitoringdashboard
pymongo
flask-cors